Add logistic regression model training script and evaluate model function
